adapt ClassificationInterpretation to work for all data types
